Test each cBPF insn for C and eBPF
Previously we only checked that each instruction could be converted to
C, but not the actual behavior of converted instructions.

Add helpers to convert C and eBPF to loadable eBPF programs that we can
prog_test.

Add a test for every cBPF instruction that attempts to verify it behaves
as expected after being converted.

Running the tests can be slow (due to calling out to clang for the C
backend); run them in parallel.