C: 0 initialize a & x, check for unintialized m[]
cBPF programs start with a, and x zero initialized, but we didn't ensure
this in our C output. Using Clang with -WAll -WError on the generated
output would have generated an error.

I'm not sure whether m[] should be 0 initialized or not. The kernel's
cBPF to eBPF conversion doesn't 0 initialize it, but doesn't check that
no uninitialized loads are present in the cBPF.
As a workaround, detected uninitialized uses of m[], and reject those
programs.