Insert less packet guards
Instead of considering only the packet accesses made in the current
block, take the accesses of target / children blocks into account.

With typical libpcap programs of the form:

load 3
if != foo no_match

load 5
if != bar no_match

The first guard can safely be increased to 5 without changing the result
of the program.