Add basic examples
Add basic examples that show how to use cbpfc based on the test suite
that builds programs to test ToC() and ToEBPF().

Add a separate helper in each example that is used by the example
function, as well as the tests. This reduces code duplication, and
ensures the examples work properly - the generated programs are tested.

It's not clear what we could usefully print in the examples to use with:

    // Output:

So the example functions are only compiled, not run. They're both only
call the helper functions though, so I think this is ok.