Expose cBPF filter return value
Expose the actual cBPF return value, instead of just match / no-match.
This allows users to make use of it, for example to only expose that
many bytes of the matching packet using perf.

This simplifies using the ebpf target as well, as the match / nomatch
blocks could previously end up as dead code. There is now a single
ResultLabel, which can't be dead code.