Replace `block.insert` with a new slice
`block.insert` inserts a new instruction in an existing block. This
isn't very nice, as it means we can't use normal `for pc, insn := range
block.insns` for loops to iterate over the insns and insert new ones at
the same time. This leads to us manually advancing `pc`, and using
`pc++` to skip over the new instruction we've inserted.

Rework the users of block.insert to create a slice of the new
instructions, and replace block.insns with it.