tests: Run end to end tests in subtests, replace XDPAction
Currently the end to end tests support checking both the CBPF to C to
EBPF and CBPF to EBPF backends.

Run these in separate subtests so that we can easily check the result of
either, including if both fail.
This introduces a "backend" abstraction to allow us to delay compiling /
doing anything until we are in the subtest.

Have the backend return a result type expressing if the packet matched
the filter, as the XDPDrop == packet matched is an arbitrary construct
that doesn't map well to other potential backends, such as a native cBPF
one.