tests: Use unix socket to check kernel behavior
The end to end tests check that the eBPF and C backends work the same
way, and produce the expected result. But they do not check that a cBPF
filter loaded into the kernel behaves the same.

Add a new test backend that uses a Unix socket to ensure that the kernel
behaves identically. This will allow us to add future tests for corner
cases (notably 32 bit ALU operations).

Skip the ZeroInitScratch test for now, as the kernel seems to reject such
filters. Tracked in #9.