Drop uninitialized scratch register support
When adding the kernel tests, we discovered that the kernel does not
accept cBPF programs that use scratch registers uninitialized.

Remove the logic we have to zero initialize uninitialized scratch
registers.

Rework some of the tests to check for uninitialized registers instead,
as we seem to be a bit light on such tests.

Fixes #9.