Support jumps with 0 skips
We try to be clever and detect when a block falls through to
another block, versus when it is explicitly jumped to.
This allows us to only generate goto labels in C for explicit jumps, and
not fall throughs.
Fall throughs are detected using 0 value skips. Legitimate Jump
instructions with 0 skips are also detected as fall throughs, causing
the label to be omitted in the generated C, failing compilation.

Instead of being clever, unconditionally emit labels, even for fall
through cases. To avoid clang errors about unused labels mark all of
these labels as unused.