Reset guard for blocks ending in clobber
requiredGuard func(insns []instruction) (int, packetGuard) determines
what packet guards a block needs, and how many instructions of insns
this guard would cover: indirect guards are invalidated by writing to
RegX.

But this doesn't allow us to differentiate between a guard that covers
all of insns as is still valid afterwards and a guard that also covers
all of insns but is not valid afterwards (because the last instruction
in insns writes to RegX). This would lead us to think a guard is still
valid when it isn't.

Rework requiredGuard() to return 0 for "guard covers all insns and is
still valid". Returning len(insns) now means the guard covers all insns
but isn't valid anymore.

Fixes #20. Found by fuzzing.