Validate variable packet offsets, support negative offsets
The kernel imposes the same restrictions as constant packet loads
to variable packet load, but to the sum of RegX + instruction.Offset:

* Offsets cannot be negative.
* Offsets must be less than 2^16. While cBPF allows greater offsets (up
  to 2^31), these cannot match a packet.

This allows RegX or the constant offset to be negative, as long as their
sum is positive.

Because the verifier does not let us use packet pointers with potentially
negative offsets (see the packetGuardIndirect comment for details), we need
to track the start and end read for indirect packet guards: before we only
track the end (and we still do for absolute packet loads).

Fixes #22 for variable offsets / packet accesses. Found by fuzzing.