Replace packetGuard with interface to track more info
Currently packetGuard only tracks the end / highest byte of the packet
read for the covered instructions.

This is sufficient for absolute loads, but to correctly support indirect
loads we also need to track the start / lowest byte of the packet read
(see next commit).

Introduce a packetGuard interface so absolute and indirect loads can
continue to use the same core logic for exploring blocks, but separate
logic for combining packetGuards / determining the biggest one.