CI: Ubuntu 22.04, clang 11, 14, and 20
Github actions has dropped support for ubuntu 20.04, switch to 22.04.

This doesn't let us easily test clang-9, as it's not packaged either by
Ubuntu or apt.llvm.org. But that version is six years old, so it's not
worth the hassle of building it ourselves.

Use clang 11 (which is packaged by Ubuntu for 22.04, but not 24.04),
keep 14, and add the latest version: 20.